+++ src/ExternalSpeller.php
@@ -9,6 +9,10 @@
 
 namespace Mekras\Speller;
 
+use Mekras\Speller\Exception\EnvironmentException;
+use Mekras\Speller\Exception\ExternalProgramFailedException;
+use Mekras\Speller\Source\EncodingAwareSource;
+use Mekras\Speller\Source\Source;
 use Symfony\Component\Process\Exception\InvalidArgumentException;
 use Symfony\Component\Process\Exception\RuntimeException;
 use Symfony\Component\Process\Process;
@@ -44,6 +48,68 @@ abstract class ExternalSpeller implements Speller
     public function __construct($binaryPath)
     {
         $this->binary = (string) $binaryPath;
+    }
+
+    /**
+     * Check text.
+     *
+     * Check given text and return an array of spelling issues.
+     *
+     * @param Source $source    Text source to check.
+     * @param array  $languages List of languages used in text (IETF language tag).
+     *
+     * @return Issue[]
+     *
+     * @throws EnvironmentException
+     * @throws ExternalProgramFailedException
+     * @throws InvalidArgumentException
+     * @throws \Symfony\Component\Process\Exception\LogicException
+     *
+     * @see   http://tools.ietf.org/html/bcp47
+     * @since x.x
+     */
+    public function checkText(Source $source, array $languages)
+    {
+        $process = $this->createProcess(
+            $this->createArguments($source, $languages),
+            $this->createEnvVars($source, $languages)
+        );
+
+        /** @noinspection PhpParamsInspection */
+        $process->setInput($source->getAsString());
+        try {
+            $process->run();
+        } catch (RuntimeException $e) {
+            throw new ExternalProgramFailedException(
+                $process->getCommandLine(),
+                $e->getMessage(),
+                0,
+                $e
+            );
+        }
+
+        try {
+            $exitCode = $process->getExitCode();
+        } catch (RuntimeException $e) {
+            throw new EnvironmentException($e->getMessage(), $e->getCode(), $e);
+        }
+
+        if (0 !== $exitCode) {
+            throw new ExternalProgramFailedException(
+                $process->getCommandLine(),
+                $process->getErrorOutput(),
+                $exitCode
+            );
+        }
+
+        $output = $process->getOutput();
+
+        if ($source instanceof EncodingAwareSource
+            && strcasecmp($source->getEncoding(), 'UTF-8') !== 0) {
+            $output = iconv($source->getEncoding(), 'UTF-8', $output);
+        }
+
+        return $this->parseOutput($output);
     }
 
     /**
@@ -84,6 +150,51 @@ abstract class ExternalSpeller implements Speller
     }
 
     /**
+     * Create arguments for external speller.
+     *
+     * @param Source $source    Text source to check.
+     * @param array  $languages List of languages used in text (IETF language tag).
+     *
+     * @return string[]
+     *
+     * @since x.x
+     *
+     * @SuppressWarnings(PMD.UnusedFormalParameter)
+     */
+    protected function createArguments(Source $source, array $languages)
+    {
+        return [];
+    }
+
+    /**
+     * Create environment variables for external speller.
+     *
+     * @param Source $source    Text source to check.
+     * @param array  $languages List of languages used in text (IETF language tag).
+     *
+     * @return string[]
+     *
+     * @since x.x
+     *
+     * @SuppressWarnings(PMD.UnusedFormalParameter)
+     */
+    protected function createEnvVars(Source $source, array $languages)
+    {
+        return [];
+    }
+
+    /**
+     * Parse external speller output.
+     *
+     * @param string $output
+     *
+     * @return Issue[]
+     *
+     * @since x.x
+     */
+    abstract protected function parseOutput($output);
+
+    /**
      * Create new instance of external program.
      *
      * @param string|string[]|null $args Command arguments.
@@ -91,13 +202,17 @@ abstract class ExternalSpeller implements Speller
      *
      * @return Process
      *
+     * @throws ExternalProgramFailedException
      * @throws InvalidArgumentException
-     * @throws RuntimeException
      */
     protected function createProcess($args = null, array $env = [])
     {
         $command = $this->composeCommand($args, $env);
-        $process = new Process($command);
+        try {
+            $process = new Process($command);
+        } catch (RuntimeException $e) {
+            throw new ExternalProgramFailedException($command, $e->getMessage(), 0, $e);
+        }
         $process->setTimeout($this->timeout);
 
         return $process;