+++ tests/Ispell/fixtures/bin/ispell.sh
@@ -1,0 +1,13 @@
+#!/bin/sh
+#
+# ispell binary stub
+#
+
+folder=$(dirname $0)
+
+case "$*" in
+    *'-a'*)
+        cat "$folder/../check.txt"
+        ;;
+esac
+exit 0